blood

+++ solo/data/classification_dataloader.py
@@ -258,7 +258,7 @@ def prepare_datasets(
             size=128
         )
         for i in range(len(train_dataset[0][1])):
-            train_dataset[0][1][i] = train_dataset[0][1][i].astype(np.int32)
+            train_dataset[0][1][i] = train_dataset[0][1][i].astype(np.float16)
         DatasetClass = vars(medmnist)[dataset]
         val_dataset = DatasetClass(
             root=val_data_path,
@@ -268,7 +268,7 @@ def prepare_datasets(
             size=128
         )
         for i in range(len(train_dataset[0][1])):
-            val_dataset[0][1][i] = val_dataset[0][1][i].astype(np.int32)
+            val_dataset[0][1][i] = val_dataset[0][1][i].astype(np.float16)
 
     elif dataset == "stl10":
         train_dataset = STL10(
